fix: add widget title back, clear widget when it disappears, and move debug to its own object
