+++ src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/events/info/TabWidgetChangeEvent.kt
@@ -3,6 +3,7 @@ package tech.thatgravyboat.skyblockapi.api.events.info
 import net.minecraft.network.chat.Component
 import tech.thatgravyboat.skyblockapi.api.events.base.EventBus
 import tech.thatgravyboat.skyblockapi.api.events.base.SkyBlockEvent
+import tech.thatgravyboat.skyblockapi.impl.events.TabListEventHandler
 import tech.thatgravyboat.skyblockapi.utils.extentions.toFormattedName
 
 data class TabWidgetChangeEvent(
@@ -61,6 +62,12 @@ enum class TabWidget {
     MOONGLADE_BEACON,
     FIRE_SALE,
     ;
+
+    val currentLines: List<String>
+        get() = TabListEventHandler.widgets[this].orEmpty()
+
+    val isActive: Boolean
+        get() = this in TabListEventHandler.widgets
 
     private val string = toFormattedName()
